+++ TPINT_GRUPO_6_LAB4_Version1/TPINT_GRUPO_6_LAB4/src/negocioImpl/ClienteNegocioImpl.java
@@ -51,7 +51,7 @@ public class ClienteNegocioImpl implements ClienteNegocio {
 	}
 
 	@Override
-	public boolean ValidacionCuil(int cuil) {
+	public boolean ValidacionCuil(long cuil) {
 		return clienteDao.ValidacionCuil(cuil);
 	}
 
@@ -98,7 +98,7 @@ public class ClienteNegocioImpl implements ClienteNegocio {
 	}
 
 	@Override
-	public boolean ValidacionCuilModificar(int cuil, int id) {
+	public boolean ValidacionCuilModificar(long cuil, int id) {
 		
 		return clienteDao.ValidacionCuilModificar(cuil, id);
 	}
